updated gitinore

+++ src/main/java/Logic/TarifAnolizer.java
@@ -13,10 +13,10 @@ import static java.util.stream.Collectors.toMap;
 public class TarifAnolizer {
 
 
-    public static TreeMap<Double, Article> execute(Map<Double, Article> old, Map<Double, NewTarif> tarifs) {
-        TreeMap<Double, Article> result = new TreeMap<>();
+    public static TreeMap<Integer, Article> execute(Map<Integer, Article> old, Map<Integer, NewTarif> tarifs) {
+        TreeMap<Integer, Article> result = new TreeMap<>();
 
-        for (Double ean : tarifs.keySet()) {
+        for (Integer ean : tarifs.keySet()) {
             if ((old.get(ean) != null) &&
                     (!old.get(ean).getPrice().equals(tarifs.get(ean).getPrice())) &&
                     (old.get(ean).getRemaings() != 0)
@@ -27,7 +27,7 @@ public class TarifAnolizer {
 
             }
         }
-        for (Double f: PromoRepo.getInstance().getAcessToRepository().keySet())
+        for (Integer f: PromoRepo.getInstance().getAcessToRepository().keySet())
               {
                   result.remove(f);
 
@@ -35,14 +35,14 @@ public class TarifAnolizer {
         return result;
     }
 
-    public static TreeMap<Double, Article> executePromo(Map<Double, Article> old, Map<Double, PromoTarif> tarifs) {
-        TreeMap<Double, Article> result = new TreeMap<>();
+    public static TreeMap<Integer, Article> executePromo(Map<Integer, Article> old, Map<Integer, PromoTarif> tarifs) {
+        TreeMap<Integer, Article> result = new TreeMap<>();
         Date now = new Date();
         int nDay = now.getDay();
         int nMonth = now.getMonth();
         int nYear = now.getYear();
         Date compare = new Date(now.getYear(), now.getMonth(), now.getDay());
-        for (Double ean : tarifs.keySet()) {
+        for (Integer ean : tarifs.keySet()) {
             if (old.get(ean) != null) {
 
                 int tDay = tarifs.get(ean).getStartTarif().getDay();
@@ -60,15 +60,15 @@ public class TarifAnolizer {
         return result;
     }
 
-    public static TreeMap<Double, Article> executePromoGettingOld(Map<Double, Article> old, Map<Double, PromoTarif> tarifs) {
-        TreeMap<Double, Article> result = new TreeMap<>();
+    public static TreeMap<Integer, Article> executePromoGettingOld(Map<Integer, Article> old, Map<Integer, PromoTarif> tarifs) {
+        TreeMap<Integer, Article> result = new TreeMap<>();
         Date now = new Date();
         int nDay = now.getDay();
         int nMonth = now.getMonth();
         int nYear = now.getYear();
         Date compare = new Date(now.getYear(), now.getMonth(), now.getDay());
 
-        for (Double ean : tarifs.keySet()) {
+        for (Integer ean : tarifs.keySet()) {
             if (old.get(ean) != null) {
 
                 int tDay = tarifs.get(ean).getEndTarif().getDay();
@@ -86,9 +86,9 @@ public class TarifAnolizer {
         return result;
     }
 
-    public static TreeMap<Double, Article> executePromoEndedYesterday(Map<Double, Article> old, Map<Double, PromoTarif> tarifs) {
-        TreeMap<Double, Article> result = new TreeMap<>();
-        for (Double ean : tarifs.keySet()) {
+    public static TreeMap<Integer, Article> executePromoEndedYesterday(Map<Integer, Article> old, Map<Integer, PromoTarif> tarifs) {
+        TreeMap<Integer, Article> result = new TreeMap<>();
+        for (Integer ean : tarifs.keySet()) {
             if (old.get(ean) != null) {
                 if  (old.get(ean).getRemaings() != 0) {
                     old.get(ean).setNewPrice(tarifs.get(ean).getPrice());
@@ -97,7 +97,7 @@ public class TarifAnolizer {
             }
         }
 
-        for (Double f: PromoRepo.getInstance().getAcessToRepository().keySet())
+        for (Integer f: PromoRepo.getInstance().getAcessToRepository().keySet())
         {if ((tarifs.get(f)!=null) && (PromoRepo.getInstance().getAcessToRepository().get(f)!=null)) {
 
             if (tarifs.get(f).getPrice().equals(PromoRepo.getInstance().getAcessToRepository().get(f).getPrice()))
