+++ src/main/java/Repositories/ArticleReposNew.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class ArticleReposNew {
-    static Map<Double, Article> reposSet;
+    static Map<Integer, Article> reposSet;
     private static ArticleReposNew ourInstance = new ArticleReposNew();
 
     private ArticleReposNew() {
@@ -21,11 +21,11 @@ public class ArticleReposNew {
         reposSet.put(article.getEan(), article);
     }
 
-    public Article getArticleByEan(Double ean) {
+    public Article getArticleByEan(Integer ean) {
         return reposSet.get(ean);
     }
 
-    public Map<Double, Article> getRepoAcess() {
+    public Map<Integer, Article> getRepoAcess() {
         return reposSet;
     }
 }
